deps: add support for BoringSSL QUIC APIs
Sourced from:
https://github.com/akamai/openssl/commits/OpenSSL_1_1_1d-quic

The APIs to support the QUIC implementation are a port of the BoringSSL
implementation that has not yet landed in OpenSSL. They must be
re-applied
separately after updating the openssl source as described above. The
current patch implementation can be found in the `deps/openssl/patches`
directory in the file
`0001-deps-add-support-for-BoringSSL-QUIC-APIs.patch`.

```text
$ git am deps/openssl/patches 0001-deps-add-support-for-BoringSSL-QUIC-APIs.patch
```

The patch file itself is generated by squashing commits from the
`OpenSSL_1_1_1d-quic` fork of the Akamai OpenSSL fork located
[here](https://github.com/akamai/openssl), starting with
[this commit](https://github.com/akamai/openssl/commit/f910151a5b60eb7b90d274332368226cc67479df),
then applying additional edits to update the implementation to
openssl-1.1.1e. As OpenSSL updates are made, additional updates
to the patch may be necessary to keep the patch in sync.

The patch is currently supported only for openssl-1.1.1e.