tools: update QuicSocket link in type-parser.js
Currently the following error is reported by the doc make target:
/quic/quic/tools/doc/allhtml.js:88
  if (!ids.has(match[1])) throw new Error(`link not found: ${match[1]}`);
                          ^

Error: link not found: quic_quic_createsocket_options
 at Object.<anonymous> (/quic/quic/tools/doc/allhtml.js:88:33)
    ...
make[1]: *** [Makefile:807: out/doc/api/all.html] Error 1
make: *** [Makefile:756: doc-only] Error 2

This commit updates the QuicSocket link to point to the
net.createQuicSocket section. I was not sure if this should really point
to the class but since was what it pointed to before I opted not to
change it.

PR-URL: https://github.com/nodejs/quic/pull/364
Reviewed-By: James M Snell <jasnell@gmail.com>
Reviewed-By: Anna Henningsen <anna@addaleax.net>