quic: fix module_categories cannot_be_required set
Currently, when configuring without quic support the code-cache test
will fail. This is because the quic js files in lib/internal/quic are
being added to the set of modules that cannot be required. When quic
support is enabled NODE_EXPERIMENTAL_QUIC will be set in node.gyp.

This commit updates the macro check to only check that
NODE_EXPERIMENTAL_QUIC is defined.

PR-URL: https://github.com/nodejs/quic/pull/375
Reviewed-By: James M Snell <jasnell@gmail.com>